global: improvements
