+++ examples/app.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2015 CERN.
+# Copyright (C) 2015, 2016 CERN.
 #
 # Invenio is free software; you can redistribute it
 # and/or modify it under the terms of the GNU General Public License as
@@ -22,15 +22,30 @@
 # waive the privileges and immunities granted to it by virtue of its status
 # as an Intergovernmental Organization or submit itself to any jurisdiction.
 
+r"""Minimal Flask application example for development.
 
-"""Minimal Flask application example for development.
-
-Run example development server:
+Create database and tables:
 
 .. code-block:: console
 
-   $ cd examples
-   $ python app.py
+    $ cd examples
+    $ flask -a app.py db init
+    $ flask -a app.py db create
+
+Load demo records from invenio-records (see
+invenio_records/data/marc21/bibliographic.xml):
+
+.. code-block:: console
+
+    $ dojson do -i data/marc21/bibliographic.xml -l marcxml marc21 | \
+        flask -a app.py records create
+
+Mint the records:
+
+.. code-block:: console
+
+    $ flask -a app.py fixtures oaiserver
+
 """
 
 from __future__ import absolute_import, print_function
@@ -41,9 +56,13 @@ from flask import Flask
 from flask_cli import FlaskCLI
 from invenio_assets import InvenioAssets
 from invenio_db import InvenioDB
+from invenio_pidstore.minters import recid_minter
 from invenio_records import InvenioRecords
+from invenio_records.models import RecordMetadata
+from sqlalchemy.orm.attributes import flag_modified
 
 from invenio_oaiserver import InvenioOAIServer
+from invenio_oaiserver.minters import oaiid_minter
 
 # Create Flask application
 app = Flask(__name__)
@@ -71,9 +90,12 @@ def oaiserver():
     from invenio_oaiserver.models import OAISet
     from invenio_records.api import Record
 
-    db.session.add(OAISet(spec='test', name='Test', description="test desc"))
+    # create a OAI Set
+    db.session.add(OAISet(spec='test', name='Test', description="test desc",
+                          search_pattern="title:Test0"))
     db.session.commit()
 
+    # create a record
     schema = {
         'type': 'object',
         'properties': {
@@ -83,4 +105,13 @@ def oaiserver():
         'required': ['title'],
     }
     Record.create({'title': 'Test0', '$schema': schema})
+    db.session.commit()
+
+    # mint all records
+    records = RecordMetadata.query.all()
+    for record in records:
+        recid_minter(record.id, record.json)
+        oaiid_minter(record.id, record.json)
+        flag_modified(record.model, 'json')
+
     db.session.commit()
